移除非测试环境下对 JUnit 的依赖
